Qwen3-TTS/pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel", "ninja"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen-tts"
version = "0.1.1"
description = "Qwen3-TTS python package"
readme = "README.md"
requires-python = ">=3.9,<3.13"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
license = { text = "Apache-2.0" }
authors = [{ name = "Alibaba Qwen Team" }]

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "websockets",
  "librosa",
  "scipy",
  "numpy",
  "torch",
  "torchaudio",
  "soundfile",
  "huggingface-hub",
  "transformers",
  "accelerate",
  "einops",
  "sox",
  "onnxruntime",
  "datasets",
  "torchcodec",
  "python-dotenv",
  "tqdm",
  "gradio",
]

[project.urls]
Homepage = "https://github.com/Qwen/Qwen3-TTS"
Repository = "https://github.com/Qwen/Qwen3-TTS"

[project.scripts]
qwen-tts-demo = "qwen_tts.cli.demo:main"

[tool.setuptools]
packages = { find = { where = ["."] , include = ["qwen_tts*"] } }
include-package-data = true

[tool.setuptools.package-data]
qwen_tts = ["py.typed", "**/*.npz"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
]
